requirements.txt
PyJWT==2.8.0
pytest==8.1.1
pytest-cov==5.0.0
